tweak(functions): more protections

+++ src/ft_putendl_fd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putendl_fd.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mm <mm@student.42.fr>                      +#+  +:+       +#+        */
+/*   By: martiper <martiper@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/10 14:03:21 by mm                #+#    #+#             */
-/*   Updated: 2023/04/13 02:01:29 by mm               ###   ########.fr       */
+/*   Updated: 2023/05/17 09:12:14 by martiper         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,6 @@
 
 void	ft_putendl_fd(char *s, int fd)
 {
-	ft_putstr_fd(s, (unsigned int)fd);
-	ft_putchar_fd('\n', (unsigned int)fd);
+	ft_putstr_fd(s, fd);
+	ft_putchar_fd('\n', fd);
 }
